+++ Models/Account/Repository.php
@@ -1,0 +1,62 @@
+<?php
+/**
+ * Copyright (c) 2017, Nosto Solutions Ltd
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright notice,
+ * this list of conditions and the following disclaimer.
+ *
+ * 2. Redistributions in binary form must reproduce the above copyright notice,
+ * this list of conditions and the following disclaimer in the documentation
+ * and/or other materials provided with the distribution.
+ *
+ * 3. Neither the name of the copyright holder nor the names of its contributors
+ * may be used to endorse or promote products derived from this software without
+ * specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
+ * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+ * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+ * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+ * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+ * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ *
+ * @author Nosto Solutions Ltd <shopware@nosto.com>
+ * @copyright Copyright (c) 2016 Nosto Solutions Ltd (http://www.nosto.com)
+ * @license http://opensource.org/licenses/BSD-3-Clause BSD 3-Clause
+ */
+
+/**
+ * Class Shopware_Plugins_Frontend_NostoTagging_Models_Account_Repository
+ */
+class Shopware_Plugins_Frontend_NostoTagging_Models_Account_Repository
+{
+    /**
+     * @param \Shopware\CustomModels\Nosto\Account\Account $account
+     * @return bool
+     */
+    public function isAccountAlreadyRegistered(
+        \Shopware\CustomModels\Nosto\Account\Account $account
+    ) {
+        /** @var Shopware\CustomModels\Nosto\Account\Account $existingAccount */
+        $existingAccount = Shopware()
+            ->Models()
+            ->getRepository("Shopware\CustomModels\Nosto\Account\Account")
+            ->findOneBy(array('name' => $account->getName()));
+
+        // If an account has been found, and the shop id is different from current shop, then it means
+        // the admin is trying to map same nosto account to two sub shops. It is not allowed.
+        if ($existingAccount != null && $existingAccount->getShopId() !== $account->getShopId()) {
+            return true;
+        }
+        return false;
+    }
+}
